agents/fast-agent/src/agent/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Acuvity AI Chatbot"
authors = [
    {name = "Satyam Sinha",email = "satyam@acuvity.ai"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "alembic (>=1.12.0)",
    "descope (>=0.1.0)",
    "fast-agent-mcp @ git+https://github.com/acuvity/fast-agent.git",
    "fastapi (>=0.103.1)",
    "mcp",
    "openinference-instrumentation-mcp",
    "opentelemetry-api (>=1.31.1,<2.0.0)",
    "opentelemetry-exporter-otlp (>=1.34.0,<2.0.0)",
    "opentelemetry-instrumentation-fastapi",
    "opentelemetry-instrumentation-httpx",
    "opentelemetry-instrumentation-threading",
    "opentelemetry-sdk (>=1.34.0,<2.0.0)",
    "passlib (>=1.7.4)",
    "pydantic (>=2.4.2)",
    "psycopg2-binary (>=2.9.6)",
    "python-jose[cryptography] (>=3.3.0)",
    "sqlalchemy (>=2.0.21)",
    "uvicorn (>=0.23.2)",
]

[tool.deptry.per_rule_ignores]
DEP002 = ["psycopg2-binary"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
